add build temp and outputs to gitignore
